requirements.txt
numpy
scipy
matplotlib
pillow
scikit-learn
tqdm
joblib
opencv-python
tensorflow
keras
requests